ready for lab

+++ Lab07.Logic/Rectangle.cs
@@ -2,7 +2,7 @@ namespace Lab07.Logic;
 
 public class Rectangle : AbstractGraphic2D
 {
-    public Rectangle(int v1, int v2, int v3, int v4)
+    public Rectangle(int v1, int v2, int v3, decimal v4)
     {
     }
